Mac-friendly version of oesred.cl
Changes how lists are read in zero subtraction routine.  

Moved syntax where captial letters before @ signs (ex. T@flat.dat, ZT@flat.dat...) are replaced by copying the contents of flat.dat with T,Z, and naming the files tflat.dat and ztflat.dat respectively. 

This may be specific to an installation of IRAF on a 2019 intel-chip macbook, but if others with macs have issues with OESRED this may be a good solution.